Basic ServerHello response

+++ Crypto/IO/TLS/TlsVersion.cs
@@ -4,10 +4,10 @@
     {
         public static readonly TlsVersion TLS1_2 = new TlsVersion(3, 3);
 
-        public readonly int Major;
-        public readonly int Minor;
+        public readonly byte Major;
+        public readonly byte Minor;
 
-        public TlsVersion(int major, int minor)
+        public TlsVersion(byte major, byte minor)
         {
             Major = major;
             Minor = minor;
